Move server code into separate files
