+++ internal/ws/server.go
@@ -2,8 +2,6 @@ package ws
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"net/http"
 	"os"
 	"time"
@@ -198,155 +196,4 @@ func (s *Server) sendSessionToken(c *websocket.Conn) error {
 
 func (s *Server) ConnectionEstablished() bool {
 	return s.state.active
-}
-
-func (s *Server) sendUpdate(update *Message) error {
-	if !s.ConnectionEstablished() {
-		err := errors.New("connection was not established yet")
-		return err
-	}
-
-	if s.updates != nil {
-		select {
-		case <-s.ctx.session.ctx.Done():
-			return s.ctx.session.ctx.Err()
-		case s.updates <- update:
-			return nil
-		}
-	}
-
-	return errors.New("update channel has not been initialized")
-}
-
-func (s *Server) Send(name string, body ...interface{}) error {
-	update, err := newUpdate(name, body...)
-	if err != nil {
-		return err
-	}
-
-	return s.sendUpdate(update)
-}
-
-func (s *Server) SendError(err error) error {
-	return s.Send("error", err.Error())
-}
-
-func (s *Server) SendErrorf(format string, a ...any) error {
-	return s.SendError(fmt.Errorf(format, a...))
-}
-
-func (s *Server) processCommand(message *Message) error {
-	message.server = s
-
-	if message.Name == "ping" {
-		return message.Reply("pong")
-	} else {
-		s.broadcastEvent(s.ctx.session.ctx, message.Name, message)
-	}
-
-	return nil
-}
-
-func (s *Server) readCommand(c *websocket.Conn) {
-	typ, bytes, err := c.Read(s.ctx.session.ctx)
-	status := websocket.CloseStatus(err)
-
-	if status != -1 {
-		// Cancel session context if client sends close request.
-		log.Debugw("Got close request", "status", status.String())
-		log.Debug("Closing session")
-		s.closeSession()
-		return
-	} else if err != nil {
-		// Cancel connection context if client loses connection,
-		// leaving session context active and giving the client
-		// ability to reconnect and continue work.
-		log.Debugw(
-			"Got error while reading message",
-			"error", err,
-			"message", string(bytes),
-		)
-		log.Debug("Closing connection")
-		s.closeConnection()
-		return
-	} else {
-		if typ != websocket.MessageText {
-			log.Debug("Only text messages are allowed")
-			s.SendError(fmt.Errorf("only text messages are allowed"))
-			return
-		}
-
-		log.With("message", string(bytes)).Debug("Received message")
-
-		command, err := commandFromBytes(bytes)
-		if err != nil {
-			log.Debugw(
-				"Got error while trying to parse command",
-				"command", string(bytes),
-				"error", err,
-			)
-
-			s.SendError(err)
-
-			return
-		}
-
-		err = s.processCommand(command)
-		if err != nil {
-			log.Debugw(
-				"Got error while processing command",
-				"command", command.Name,
-				"error", err,
-			)
-
-			s.SendError(err)
-
-			return
-		}
-	}
-}
-
-func (s *Server) listenForCommands(c *websocket.Conn) {
-	for {
-		select {
-		case <-s.ctx.session.ctx.Done():
-		case <-s.ctx.connection.ctx.Done():
-			return
-		default:
-			s.readCommand(c)
-		}
-	}
-}
-
-func (s *Server) listenForUpdates(c *websocket.Conn) {
-	for {
-		select {
-		case <-s.ctx.session.ctx.Done():
-		case <-s.ctx.connection.ctx.Done():
-			return
-		case update := <-s.updates:
-			if !s.ConnectionEstablished() {
-				log.Debug("Handshake was not established yet")
-				continue
-			}
-
-			bytes, err := update.Bytes()
-			if err != nil {
-				log.Debugf(
-					"Got error while trying to send update '%v': %s",
-					string(bytes), err,
-				)
-				continue
-			}
-
-			err = c.Write(s.ctx.connection.ctx, websocket.MessageText, bytes)
-			if err != nil {
-				log.Debugf(
-					"Got error while trying to send update '%v': %s",
-					string(bytes), err,
-				)
-				continue
-			}
-		}
-	}
 }
